# Optimize PRBAC SQL queries