--- a/corehq/apps/accounting/utils.py
+++ b/corehq/apps/accounting/utils.py
@@ -77,7 +77,7 @@ def fmt_product_rate_dict(product, product_rate=None):
 
 
 def get_privileges(plan_version):
-    role = plan_version.role
+    role = plan_version.role.get_cached_role()
     return set([grant.to_role.slug for grant in role.memberships_granted.all()])
 
 
@@ -116,10 +116,9 @@ def domain_has_privilege(domain, privilege_slug, **assignment):
     from corehq.apps.accounting.models import Subscription
     try:
         plan_version = Subscription.get_subscribed_plan_by_domain(domain)[0]
-        roles = Role.objects.filter(slug=privilege_slug)
-        if not roles:
+        privilege = Role.get_privilege(privilege_slug, assignment)
+        if privilege is None:
             return False
-        privilege = roles[0].instantiate(assignment)
         if plan_version.role.has_privilege(privilege):
             return True
     except ProductPlanNotFoundError:
@@ -211,10 +210,9 @@ def get_customer_cards(account, username, domain):
 
 
 def is_accounting_admin(user):
-    roles = Role.objects.filter(slug=privileges.ACCOUNTING_ADMIN)
-    if not roles:
+    accounting_privilege = Role.get_privilege(privileges.ACCOUNTING_ADMIN)
+    if accounting_privilege is None:
         return False
-    accounting_privilege = roles[0].instantiate({})
     try:
         return user.prbac_role.has_privilege(accounting_privilege)
     except (AttributeError, UserRole.DoesNotExist):