--- a/corehq/apps/users/util.py
+++ b/corehq/apps/users/util.py
@@ -10,7 +10,7 @@ from corehq import privileges
 from dimagi.utils.couch.database import get_db
 from django.core.cache import cache
 from django_prbac.exceptions import PermissionDenied
-from django_prbac.utils import ensure_request_has_privilege
+from django_prbac.utils import has_privilege
 
 
 WEIRD_USER_IDS = [
@@ -152,9 +152,7 @@ def can_add_extra_mobile_workers(request):
     user_limit = request.plan.user_limit
     if user_limit == -1 or num_web_users < user_limit:
         return True
-    try:
-        ensure_request_has_privilege(request, privileges.ALLOW_EXCESS_USERS)
-    except PermissionDenied:
+    if not has_privilege(request, privileges.ALLOW_EXCESS_USERS):
         account = BillingAccount.get_account_by_domain(request.domain)
         if account is None or account.date_confirmed_extra_charges is None:
             return False