Review commits individually. Depends on https://github.com/dimagi/django-prbac/pull/14

Some things to consider:
- All roles and grants are stored in memory. However, the tables are pretty small. Here are the numbers from prod:

  - django_prbac_grant - 256 rows
  - django_prbac_role - 91 rows
  - django_prbac_userrole - 30 rows

  I think keeping them in memory is a reasonable tradeoff for the reduction it DB hits.
- This optimization focuses on the case where "role assignments" are not used. I checked, and none of our roles or grants are using assignments. However, things should continue to work if we start using assignments, but more optimization may be necessary.

Page used for testing http://localhost:8080/a/test/dashboard/project/
Profiling numbers come from django toolbar

Before:

  - Cold cache  
    CPU Time: 15335.91 ms (maybe not accurate?)  
    212 SQL queries in 196.54ms
  - Hot cache  
    CPU Time: 5509.83 ms  
    47 SQL queries in 44.86ms

After:

  - Cold cache  
    CPU Time: 6499.19 ms  
    6 SQL queries in 11.87ms
  - Hot cache  
    CPU Time: 4085.39 ms  
    4 SQL queries in 8.66 ms


@esoergel @biyeun cc @sravfeyn