@czue 2015-07-23
just trying to wrap my head around the caching implications. in the following scenario:

1. I don't have permission to use cloudcare
2. I change my subscription to one that has permission
3. I load cloudcare

Does caching kick in at all? I think it doesn't, and the scenario it would, would be if we decided the enterprise plan should no longer have access to cloudcare. Just want to make sure that's right. If it is the first case I would advocate for clearing them when plans change.