@millerdev 2015-07-23
@czue wrote:
> 1. I don't have permission to use cloudcare
> 2. I change my subscription to one that has permission
> 3. I load cloudcare
> 
> Does caching kick in at all? I think it doesn't

That's correct. The level of caching done in this PR will not affect that case since [a fresh copy of the user's role is looked up on each request](https://github.com/dimagi/commcare-hq/blob/abcd219fe4d88e974e971220b7da244bc98ec0d1/corehq/apps/domain/middleware.py#L36-L37). The caching part comes in when calculating if the user's role has permission to access cloudcare. At that point the user's role is [swapped with a cached copy](https://github.com/dimagi/django-prbac/blob/3b987dc0ac82708bb3a6e0e87e136c4ada78c930/django_prbac/models.py#L164)

> the scenario [this new caching] would [affect], would be if we decided the enterprise plan should no longer have access to cloudcare.

That's correct. If we changed enterprise to no longer have access to cloudcare, then people on the enterprise plan would have, at a maximum, the length of the cache duration (1 minute) to use cloudcare before they lost that privilege.