@esoergel 2015-07-24
Test failure looks real:
```
======================================================================
ERROR: testGetCasesCaching (corehq.apps.cloudcare.tests.test_api.CaseAPITest)
----------------------------------------------------------------------

Traceback (most recent call last):
  File "/home/travis/build/dimagi/commcare-hq/corehq/apps/cloudcare/tests/test_api.py", line 244, in testGetCasesCaching
    'use_cache': 'true',
  File "/home/travis/virtualenv/local/lib/python2.7/site-packages/django/test/client.py", line 473, in get
    response = super(Client, self).get(path, data=data, **extra)
  File "/home/travis/virtualenv/local/lib/python2.7/site-packages/django/test/client.py", line 280, in get
    return self.request(**r)
  File "/home/travis/virtualenv/local/lib/python2.7/site-packages/django/test/client.py", line 444, in request
    six.reraise(*exc_info)
  File "/home/travis/virtualenv/local/lib/python2.7/site-packages/django/core/handlers/base.py", line 105, in get_response
    response = middleware_method(request, callback, callback_args, callback_kwargs)
  File "/home/travis/virtualenv/local/lib/python2.7/site-packages/django/utils/decorators.py", line 29, in _wrapper
    return bound_func(*args, **kwargs)
  File "/home/travis/virtualenv/local/lib/python2.7/site-packages/django/views/decorators/debug.py", line 75, in sensitive_post_parameters_wrapper
    return view(request, *args, **kwargs)
  File "/home/travis/virtualenv/local/lib/python2.7/site-packages/django/utils/decorators.py", line 25, in bound_func
    return func(self, *args2, **kwargs2)
  File "/home/travis/build/dimagi/commcare-hq/corehq/apps/domain/middleware.py", line 29, in process_view
    self.apply_prbac(request)
  File "/home/travis/build/dimagi/commcare-hq/corehq/apps/domain/middleware.py", line 44, in apply_prbac
    request.role = Role.get_privilege('community_plan_v0').role

AttributeError: 'NoneType' object has no attribute 'role'
```