@millerdev 2015-07-24
Let's wait to merge this (also wait to update the submodule) until Monday. Don't think it will cause issues, but would hate for something to go wrong since I have a pretty busy weekend with getting ready for the bike trip.