@millerdev 2015-07-25
This is discouraging. I have tests that are failing on my local machine, but I can't figure out why. They're even failing on a fresh checkout of master. Therefore it's been difficult for me to figure out which failures are real issues caused by changes on this branch.