@benrudolph 2015-07-27
Looks like they've passed on travis. What tests aren't passing locally? Just checking, you have `SOUTH_TESTS_MIGRATE = False`