@millerdev 2015-07-27
Finally passed! Yay!! I think that means this is good to merge, assuming you peeps are happy with it.

`SOUTH_TESTS_MIGRATE = False` is in settings.py, and I have not overridden it, so that seems right to me. Will troubleshoot the failing tests on other channels.