@millerdev 2015-07-27
@amsagoff do we need to do anything special to measure if this changes HQ performance when this is deployed (presumably later today)? Theoretically it should affect nearly every request to HQ.