@amsagoff 2015-07-27
amazing! We need to wait a bit to get statistically significant results. I'll be doing this with all the changes people made on performance day (or trying to - so hard to control for all the different changes happening!).
I added it to the list with the deployment date set optimistically to today: https://docs.google.com/spreadsheets/d/1hMJ0PtXfxGJNE_dqqhzyocHJUfl8-guX2cnKY3SlKZ4/edit#gid=931792981